chore(server): refactor and unify users infrastructure tests
